Updated webdriver wait function to wait for presence of an element
